amp-3d-gltf: rename js files to match convention
